# swagger.yaml
openapi: 3.0.3
info:
  title: Event Booking System API
  description: |
    A REST API for managing events, bookings, and attendance tracking.

    ## Features
    - Browse and create upcoming events
    - Book tickets (single or multiple) with race-condition-safe transactions
    - Track event attendance via unique booking codes (UUID v4)
    - View all bookings per user

    ## Error Format
    All errors return a consistent JSON shape:
    ```json
    {
      "success": false,
      "status": 404,
      "error": "Event not found"
    }
    ```
    Validation errors also include an `errors` array with every broken rule.
  version: 1.1.0
  contact:
    name: API Support
    email: support@eventbooking.com
servers:
  - url: https://event-booking-api-production.up.railway.app
    description: Production server
  - url: http://localhost:3000
    description: Local development server
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        booking_id:
          type: integer
          example: 42
        user_id:
          type: integer
          example: 7
        event_id:
          type: integer
          example: 1
        quantity:
          type: integer
          minimum: 1
          maximum: 100
          default: 1
          description: Number of tickets booked
          example: 2
        event_title:
          type: string
          example: "Node.js Workshop 2025"
        unique_code:
          type: string
          format: uuid
          description: UUID v4 used for check-in
          example: "550e8400-e29b-41d4-a716-446655440000"
        booking_date:
          type: string
          format: date-time
          example: "2025-10-01T09:30:00.000Z"
...
    CreateBookingBody:
      type: object
      required:
        - user_id
        - event_id
      properties:
        user_id:
          type: integer
          minimum: 1
          maximum: 999999999
          example: 7
        event_id:
          type: integer
          minimum: 1
          maximum: 999999999
          example: 1
        quantity:
          type: integer
          minimum: 1
          maximum: 100
          default: 1
          description: Number of tickets to book (max 100 per request)
          example: 2
...
    Error400:
      type: object
      properties:
        success:
          type: boolean
          example: false
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "title is required and must be a non-empty string | date must be in the future"
        errors:
          type: array
          description: Present only for validation errors — lists every broken rule
          items:
            type: string
          example:
            - "title is required and must be a non-empty string"
            - "date must be in the future"
    Error404:
      type: object
      properties:
        success:
          type: boolean
          example: false
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Event not found"
    Error409:
      type: object
      properties:
        success:
          type: boolean
          example: false
        status:
          type: integer
          example: 409
        error:
          type: string
          example: "No tickets remaining for this event"
    Error500:
      type: object
      properties:
        success:
          type: boolean
          example: false
        status:
          type: integer
          example: 500
        error:
          type: string
          example: "Internal server error"
  responses:
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error500"
paths:
...
  /bookings:
    post:
      tags:
        - Bookings
      summary: Book one or more tickets for an event
      description: |
        Creates a booking for a user-event pair. Supports booking multiple tickets in a single
        request via the `quantity` field (default = 1, max = 100).

        This endpoint runs inside a MySQL transaction with row-level lock (`SELECT ... FOR UPDATE`)
        to prevent overselling under concurrent load.

        Business rules enforced:
        - Event must exist
        - `remaining_tickets >= quantity`
        - User must exist
        - User cannot book the same event twice (update not supported yet)
        - `remaining_tickets` is decremented by `quantity`
        - A UUID v4 `unique_code` is generated — use this for check-in
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
            examples:
              singleTicket:
                summary: Book 1 ticket (default)
                value:
                  user_id: 7
                  event_id: 1
              multipleTickets:
                summary: Book 3 tickets
                value:
                  user_id: 7
                  event_id: 1
                  quantity: 3
      responses:
        "201":
          description: Booking confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Booking confirmed for 2 ticket(s)"
                  data:
                    $ref: "#/components/schemas/Booking"
              examples:
                single:
                  summary: Single ticket
                  value:
                    success: true
                    message: "Booking confirmed"
                    data:
                      booking_id: 42
                      user_id: 7
                      event_id: 1
                      quantity: 1
                      event_title: "Node.js Workshop 2025"
                      unique_code: "550e8400-e29b-41d4-a716-446655440000"
                      booking_date: "2025-10-01T09:30:00.000Z"
                multiple:
                  summary: Multiple tickets
                  value:
                    success: true
                    message: "Booking confirmed for 3 ticket(s)"
                    data:
                      booking_id: 43
                      user_id: 7
                      event_id: 1
                      quantity: 3
                      event_title: "Node.js Workshop 2025"
                      unique_code: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                      booking_date: "2025-10-01T09:35:00.000Z"
        "400":
          description: Invalid request body (missing fields, non-integer IDs, quantity out of range)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
              examples:
                invalidUserId:
                  summary: user_id not a positive integer
                  value:
                    success: false
                    status: 400
                    error: "user_id must be a positive integer"
                    errors:
                      - "user_id must be a positive integer"
                quantityTooHigh:
                  summary: Quantity exceeds maximum
                  value:
                    success: false
                    status: 400
                    error: "quantity cannot exceed 100 per booking"
                    errors:
                      - "quantity cannot exceed 100 per booking"
        "404":
          description: User or event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error404"
              examples:
                eventNotFound:
                  summary: Event does not exist
                  value:
                    success: false
                    status: 404
                    error: "Event not found"
                userNotFound:
                  summary: User does not exist
                  value:
                    success: false
                    status: 404
                    error: "User not found"
        "409":
          description: Conflict — sold out, duplicate booking, or insufficient tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error409"
              examples:
                soldOut:
                  summary: No tickets remaining
                  value:
                    success: false
                    status: 409
                    error: "No tickets remaining for this event"
                insufficient:
                  summary: Requested quantity exceeds remaining tickets
                  value:
                    success: false
                    status: 409
                    error: "Only 2 tickets available"
                duplicate:
                  summary: User already booked this event
                  value:
                    success: false
                    status: 409
                    error: "User has already booked this event (update quantity not supported yet)"
        "500":
          $ref: "#/components/responses/InternalError"
...